requirements.txt
PyChromecast
spotify-token
spotipy
nfcpy
